user模块下的permission的一些功能（user用户管理和role角色以及这次写的permission接口均未调试）-等待远端nacos稳定后开始调试 （也对DTO类的一些完善）

+++ huashui-campus-life/huashui-user/src/main/java/com/huashui/user/domain/vo/RoleDetailVO.java
@@ -23,7 +23,7 @@ public class RoleDetailVO implements Serializable {
     private List<Long> permissionIds;
 
     @Schema(description = "拥有此角色的用户数")
-    private Integer userCount;
+    private Long userCount;
 
     @Schema(description = "角色ID")
     private Long id;
